Change ArchUnit kotlin package and move DefaultProblemsReportCreator.kt
Use separate parameter for report context and report file name.
